Lesson 04 (#8)
* Navigation Provider added

NavigationProvider, and the necessary functions to register lists into a top level array.

* Down Arrow functionality

Allows moving to the first child of an element when its sublist is open.
Allows to move from the last child of an open list to its parent's sibling if available, or to its default behavior.

* Complete Down Arrow functionality

SubNavigation and NavigationLink

* Testing Coverage

100%

---------

Co-authored-by: sandraclarkdev <58484559+sandraclarkdev@users.noreply.github.com>