Lesson 5 (#11)
* Tabbing

Last tab on a closed submenu button goes to the next parent in the DOM of the last item in the list (This sends a closed nav to the next sibling on the top row or out of the component altogether when the tab key is pushed when on the last closed submenu

* Closed SubMenu tabs to it's last ele,ment's next in DOM

When a submenu button has a closed sub nav list and a Tab is detected, the next focusable item is the next item in the DOM from the last element in the subnavigation. This allows tabbing into and out of the component.

* Add Escape Key functionality

* Tab out of component

Tab out of the component when on the last top button or the last top link. Close any open submenus when leaving the last link not on the top row.

* Right Arrow on Top Row

The right arrow on the top row is implemented to always move to the next item on the top row, regardless of the state of the menu underneath it.

* Click away listener added

Everything provisionally moved into Navigation Provider. Cleanup next

* Test Cleanup

linting issues.

* Refactor

Selectors from provider to hook.

* Code Cleanup

* Istanbul ignores

Setting istanbul ignores for else statements and code necessary for lesson 6

---------

Co-authored-by: sandraclarkdev <58484559+sandraclarkdev@users.noreply.github.com>